+++ src/main/java/com/hunllefhelper/ui/components/Button.java
@@ -26,4 +26,10 @@ public class Button extends JButton
 			}
 		});
 	}
+
+	public void resetStyling()
+	{
+		setEnabled(false);
+		setEnabled(true);
+	}
 }